+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Change categories:
 - utils: Introduce `pldm_edac_crc32()`
 - utils: Introduce `pldm_edac_crc8()`
 - pdr: Add pldm_pdr_delete_by_effecter_id() API
+- platform: Add encode req for GetPDRRepositoryInfo
 
 - oem: ibm: Add boot side rename state set and enum